+++ app/src/main/java/com/example/programacion_dirigida_por_eventos_aplicacion_de_gestion_de_residuos/activity/StatsActivity.java
@@ -1,0 +1,2 @@
+package com.example.programacion_dirigida_por_eventos_aplicacion_de_gestion_de_residuos.activity;public class StatsActivity {
+}
