+++ app/src/main/res/drawable/ic_calendar.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#000000"
+        android:pathData="M19,3h-1V1h-2v2H8V1H6v2H5c-1.1,0 -1.99,0.9 -1.99,2L3,21c0,1.1 0.9,2 2,2h14c1.1,0 2-0.9 2-2V5c0-1.1 -0.9-2 -2-2zM19,21H5V8h14v13zM7,10h5v5H7z" />
+</vector>
