+++ app/src/main/res/drawable/ic_stats.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#000000"
+        android:pathData="M4,10h3v10H4V10zM10.5,6h3v14h-3V6zM17,13h3v7h-3V13z" />
+</vector>
