+++ app/src/main/res/menu/drawer_menu.xml
@@ -1,0 +1,14 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/nav_calendar"
+        android:title="Calendario de recolección"
+        android:icon="@drawable/ic_calendar" />
+    <item
+        android:id="@+id/nav_map"
+        android:title="Mapa de puntos de reciclaje"
+        android:icon="@drawable/ic_map" />
+    <item
+        android:id="@+id/nav_stats"
+        android:title="Estadísticas de reciclaje"
+        android:icon="@drawable/ic_stats" />
+</menu>
